entity list is working with extracted filter logic TOOD fix tests
