+++ src/app/core/entity-components/entity-list/entity-list.component.html
@@ -39,19 +39,14 @@
 
   <div class="flex-row gap-regular flex-wrap">
     <div *ngTemplateOutlet="filterDialog"></div>
-
-    <app-list-filter
-      *ngFor="let filter of filterSelections"
-      [filterConfig]="filter.filterSettings"
-      [(selectedOption)]="filter.selectedOption"
-      [displayAsToggle]="filter.display !== 'dropdown'"
-      (selectedOptionChange)="filterOptionSelected(filter, $event)"
-      angulartics2On="click"
-      [angularticsCategory]="entityConstructor?.ENTITY_TYPE"
-      angularticsAction="list_filter"
-      [angularticsLabel]="filter.filterSettings.label"
-    >
-    </app-list-filter>
+    <app-filter
+      class="flex-row gap-regular flex-wrap"
+      [filterConfig]="filtersConfig"
+      [entityType]="entityConstructor"
+      [entities]="allEntities"
+      [withUrl]="true"
+      [(filterObj)]="filterObj"
+    ></app-filter>
   </div>
 
   <div>
@@ -60,7 +55,6 @@
         <mat-tab
           *ngFor="let item of columnGroups"
           [label]="item.name"
-          (click)="columnGroupClick(item.name)"
           angulartics2On="click"
           [angularticsCategory]="entityConstructor?.ENTITY_TYPE"
           angularticsAction="list_column_view"
