+++ src/app/core/entity-components/entity-list/entity-list.component.ts
@@ -19,10 +19,8 @@ import { Entity, EntityConstructor } from "../../entity/model/entity";
 import { FormFieldConfig } from "../entity-form/entity-form/FormConfig";
 import { EntitySubrecordComponent } from "../entity-subrecord/entity-subrecord/entity-subrecord.component";
 import { FilterGeneratorService } from "./filter-generator.service";
-import { FilterComponentSettings } from "./filter-component.settings";
 import { entityFilterPredicate } from "./filter-predicate";
 import { MatDialog } from "@angular/material/dialog";
-import { FilterOverlayComponent } from "./filter-overlay/filter-overlay.component";
 import { AnalyticsService } from "../../analytics/analytics.service";
 import { RouteTarget } from "../../../app.routing";
 import { RouteData } from "../../view/dynamic-routing/view-config.interface";
@@ -31,6 +29,7 @@ import { EntityRegistry } from "../../entity/database-entity.decorator";
 import { ScreenWidthObserver } from "../../../utils/media/screen-size-observer.service";
 import { UntilDestroy, untilDestroyed } from "@ngneat/until-destroy";
 import { DataFilter } from "../entity-subrecord/entity-subrecord/entity-subrecord-config";
+import { FilterComponent } from "../../filter/filter/filter.component";
 
 /**
  * This component allows to create a full blown table with pagination, filtering, searching and grouping.
@@ -59,6 +58,7 @@ export class EntityListComponent<T extends Entity>
   @Output() addNewClick = new EventEmitter();
 
   @ViewChild(EntitySubrecordComponent) entityTable: EntitySubrecordComponent<T>;
+  @ViewChild(FilterComponent) private filterComponent: FilterComponent<T>;
 
   isDesktop: boolean;
 
@@ -71,7 +71,6 @@ export class EntityListComponent<T extends Entity>
 
   columnsToDisplay: string[] = [];
 
-  filterSelections: FilterComponentSettings<T>[] = [];
   filterObj: DataFilter<T>;
   filterString = "";
 
@@ -133,9 +132,6 @@ export class EntityListComponent<T extends Entity>
 
         this.isDesktop = isDesktop;
       });
-    this.activatedRoute.queryParams.subscribe((params) => {
-      this.loadUrlParams(params);
-    });
   }
 
   private async buildComponentFromConfig(data: RouteData<EntityListConfig>) {
@@ -178,10 +174,6 @@ export class EntityListComponent<T extends Entity>
         this.displayColumnGroupByName(this.mobileColumnGroup);
       }
     }
-    if (changes.hasOwnProperty("allEntities")) {
-      await this.initFilterSelections();
-      this.applyFilterSelections();
-    }
     this.loadUrlParams();
   }
 
@@ -220,25 +212,11 @@ export class EntityListComponent<T extends Entity>
     }
   }
 
-  private loadUrlParams(parameters?: Params) {
-    const params = parameters || this.activatedRoute.snapshot.queryParams;
+  private loadUrlParams() {
+    const params = this.activatedRoute.snapshot.queryParams;
     if (params["view"]) {
       this.displayColumnGroupByName(params["view"]);
     }
-    this.filterSelections.forEach((f) => {
-      if (params.hasOwnProperty(f.filterSettings.name)) {
-        f.selectedOption = params[f.filterSettings.name];
-      }
-    });
-    this.applyFilterSelections();
-    if (params["search"]) {
-      this.applyFilter(params["search"]);
-    }
-  }
-
-  columnGroupClick(columnGroupName: string) {
-    this.displayColumnGroupByName(columnGroupName);
-    this.updateUrl("view", columnGroupName);
   }
 
   applyFilter(filterValue: string) {
@@ -249,44 +227,6 @@ export class EntityListComponent<T extends Entity>
     this.analyticsService.eventTrack("list_filter_freetext", {
       category: this.entityConstructor?.ENTITY_TYPE,
     });
-  }
-
-  filterOptionSelected(
-    filter: FilterComponentSettings<T>,
-    selectedOption: string
-  ) {
-    filter.selectedOption = selectedOption;
-    this.applyFilterSelections();
-    this.updateUrl(filter.filterSettings.name, selectedOption);
-  }
-
-  private applyFilterSelections() {
-    this.filterObj = this.filterSelections.reduce(
-      (obj, filter) =>
-        Object.assign(
-          obj,
-          filter.filterSettings.getFilter(filter.selectedOption)
-        ),
-      {} as DataFilter<T>
-    );
-  }
-
-  private updateUrl(key: string, value: string) {
-    const params = {};
-    params[key] = value;
-    this.router.navigate([], {
-      relativeTo: this.activatedRoute,
-      queryParams: params,
-      queryParamsHandling: "merge",
-    });
-  }
-
-  private async initFilterSelections(): Promise<void> {
-    this.filterSelections = await this.filterGeneratorService.generate(
-      this.filtersConfig,
-      this.entityConstructor,
-      this.allEntities
-    );
   }
 
   private displayColumnGroupByName(columnGroupName: string) {
@@ -301,21 +241,7 @@ export class EntityListComponent<T extends Entity>
     return this.columnGroups.findIndex((c) => c.name === columnGroupName);
   }
 
-  getNewRecordFactory(): () => T {
-    return () => new this.entityConstructor();
-  }
-
   openFilterOverlay() {
-    this.dialog.open(FilterOverlayComponent, {
-      data: {
-        filterSelections: this.filterSelections,
-        filterChangeCallback: (
-          filter: FilterComponentSettings<T>,
-          option: string
-        ) => {
-          this.filterOptionSelected(filter, option);
-        },
-      },
-    });
+    this.filterComponent.openFilterOverlay();
   }
 }
