+++ src/app/core/filter/list-filter/list-filter.component.spec.ts
@@ -1,21 +1,19 @@
 import { ComponentFixture, TestBed, waitForAsync } from "@angular/core/testing";
 
 import { ListFilterComponent } from "./list-filter.component";
-import { FilterSelection } from "../../../filter/filter-selection/filter-selection";
-import { EntityListModule } from "../entity-list.module";
-import { MockedTestingModule } from "../../../../utils/mocked-testing.module";
+import { FilterSelection } from "../filter-selection/filter-selection";
+import { EntityListModule } from "../../entity-components/entity-list/entity-list.module";
+import { MockedTestingModule } from "../../../utils/mocked-testing.module";
 
 describe("ListFilterComponent", () => {
   let component: ListFilterComponent<any>;
   let fixture: ComponentFixture<ListFilterComponent<any>>;
 
-  beforeEach(
-    waitForAsync(() => {
-      TestBed.configureTestingModule({
-        imports: [EntityListModule, MockedTestingModule.withState()],
-      }).compileComponents();
-    })
-  );
+  beforeEach(waitForAsync(() => {
+    TestBed.configureTestingModule({
+      imports: [EntityListModule, MockedTestingModule.withState()],
+    }).compileComponents();
+  }));
 
   beforeEach(() => {
     fixture = TestBed.createComponent(ListFilterComponent);
